Create custom item types
To make the code cleaner and more expandable in the future, each custom
item has been given its own class. This has been done in such a way that
does not effect user space however it lays the groundwork for v2.0.0
which will break user space by requiring users to create new instances
of the custom class as opposed to calling a function which creates the
instance.